+++ changelog.md
@@ -4,12 +4,23 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-# [Unreleased](https://github.com/MyPureCloud/purecloud-webrtc-sdk/compare/v3.6.0...HEAD)
+# [Unreleased](https://github.com/MyPureCloud/purecloud-webrtc-sdk/compare/v3.6.1...HEAD)
 ### Fixed
 * make sure output device(s) is available when devices change 
 
 ### Changed
 * added session ids to logging around devices
+
+# [v3.6.1](https://github.com/MyPureCloud/purecloud-webrtc-sdk/compare/v3.6.0...v3.6.1)
+### Added
+* Added ideal frameRate constraint to screenShare and video
+* Apply track constraints based on the media automatically
+
+### Changed
+* When possible, replace sender tracks directly instead of removing the track, then adding the new one
+
+### Fixed
+* Fixed the context for the logging handlers
 
 # [v3.6.0](https://github.com/MyPureCloud/purecloud-webrtc-sdk/compare/v3.5.1...v3.6.0)
 ### Added
@@ -17,9 +28,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * pending sessions can now be rejected with the `rejectPendingSession` function
 
 ### Fixed
-* When `mediaDevices` fires a `devicechange` event, the sdk will check to make sure all sessions have 
-media from devices that are still available. It will attempt to update if the media is no longer available 
-_or_ it will set the session mute if there are no devices available to switch to. 
+* When `mediaDevices` fires a `devicechange` event, the sdk will check to make sure all sessions have
+media from devices that are still available. It will attempt to update if the media is no longer available
+_or_ it will set the session mute if there are no devices available to switch to.
 
 ### Security
 * Merged [PR#138](https://github.com/MyPureCloud/purecloud-webrtc-sdk/pull/138) bumping `minimist`